feat(io): stream file payloads through decoded readers
## What

- Decode configured payloads as closeable streams instead of byte slices.
- Copy decoded streams directly to command output writers, so file-backed payloads are opened and streamed instead of buffered in memory.
- Keep the existing config format, command output behavior, and payload error handling intact.

## Why

- Large recorded command outputs should not require loading the whole file-backed fixture into memory before writing.
- Moving the stream contract into payload decoding keeps output writing focused on copying from the decoded source to the destination.

## Testing

- `go test ./internal/encoding ./internal/io -count=1` - passed
- `make build` - passed
- `make specs` - passed
- `make lint` - passed
- `make sec` - passed
- `make coverage` - passed
